why does it segfault only after 22.4k seeds specifically

+++ NoitaChestFinderCUDA/misc/worldgen_helpers.h
@@ -16,7 +16,7 @@
 #define WORLD_OFFSET_X 35
 
 #define BCSize 9
-__device__ __constant__ ulong blockedColors[BCSize] = {
+__device__ __constant__ uint blockedColors[BCSize] = {
 	0x00ac6eU, //load_pixel_scene4_alt
 	0x70d79eU, //load_gunpowderpool_01
 	0x70d79fU, //???
@@ -28,7 +28,7 @@ __device__ __constant__ ulong blockedColors[BCSize] = {
 	0xff0affU, //load_pixel_scene
 };
 
-__device__ bool contains(const ulong arr[BCSize], ulong val)
+__device__ bool contains(const uint arr[BCSize], uint val)
 {
 	for (int i = 0; i < BCSize; i++)
 		if (arr[i] == val) return true;
@@ -53,7 +53,7 @@ __device__ NollaPrng GetRNG(uint world_seed, int map_w)
 	return rng;
 }
 
-__device__ ulong getPos(const uint w, const uint s, const uint x, const uint y)
+__device__ uint getPos(const uint w, const uint s, const uint x, const uint y)
 {
 	return s * (w * y + x);
 }
@@ -63,9 +63,9 @@ __device__ void doCoalMineHax(byte* map, uint map_w, uint map_h)
 	{
 		for (int x = 0; x < map_w; x++)
 		{
-			ulong overlayPos = getPos(256, 3, x, y);
-			ulong i = getPos(map_w, 3, x, y);
-			ulong pix = createRGB(coalmine_overlay[overlayPos], coalmine_overlay[overlayPos + 1], coalmine_overlay[overlayPos + 2]);
+			uint overlayPos = getPos(256, 3, x, y);
+			uint i = getPos(map_w, 3, x, y);
+			uint pix = createRGB(coalmine_overlay[overlayPos], coalmine_overlay[overlayPos + 1], coalmine_overlay[overlayPos + 2]);
 			if (pix == 0x4000)
 			{
 				// pudy248 note: is not actually air, this is the main rock portion of the overlay
@@ -90,19 +90,19 @@ __device__ void doCoalMineHax(byte* map, uint map_w, uint map_h)
 	}
 }
 
-__device__ ulong getPixelColor(const byte* map, ulong pos)
+__device__ uint getPixelColor(const byte* map, uint pos)
 {
 	byte r = map[pos];
 	byte g = map[pos + 1];
 	byte b = map[pos + 2];
 	return createRGB(r, g, b);
 }
-__device__ ulong getPixelColor(const byte* map, const uint w, const uint x, const uint y)
+__device__ uint getPixelColor(const byte* map, const uint w, const uint x, const uint y)
 {
-	ulong pos = getPos(w, 3, x, y);
+	uint pos = getPos(w, 3, x, y);
 	return getPixelColor(map, pos);
 }
-__device__ void setPixelColor(byte* map, ulong pos, ulong color)
+__device__ void setPixelColor(byte* map, uint pos, uint color)
 {
 	byte r = ((color >> 16) & 0xff);
 	byte g = ((color >> 8) & 0xff);
@@ -111,13 +111,13 @@ __device__ void setPixelColor(byte* map, ulong pos, ulong color)
 	map[pos + 1] = g;
 	map[pos + 2] = b;
 }
-__device__ void setPixelColor(byte* map, uint w, uint x, uint y, ulong color)
+__device__ void setPixelColor(byte* map, uint w, uint x, uint y, uint color)
 {
-	ulong pos = getPos(w, 3, x, y);
+	uint pos = getPos(w, 3, x, y);
 	setPixelColor(map, pos, color);
 }
 
-__device__ void fill(byte* map, int w, int x1, int x2, int y1, int y2, ulong color)
+__device__ void fill(byte* map, int w, int x1, int x2, int y1, int y2, uint color)
 {
 	for (int x = x1; x <= x2; x++)
 	{
@@ -128,13 +128,13 @@ __device__ void fill(byte* map, int w, int x1, int x2, int y1, int y2, ulong col
 	}
 }
 
-__device__ void blockOutRooms(byte* map, uint map_w, uint map_h, ulong targetColor)
+__device__ void blockOutRooms(byte* map, uint map_w, uint map_h, uint targetColor)
 {
-	ulong posMax = map_w * map_h;
+	uint posMax = map_w * map_h;
 
-	for (ulong pos = 0; pos < posMax; pos++)
+	for (uint pos = 0; pos < posMax; pos++)
 	{
-		ulong color = getPixelColor(map, pos * 3);
+		uint color = getPixelColor(map, pos * 3);
 		// ~70% of pixels are black or white, so skip them
 		if (color == COLOR_BLACK)
 		{
@@ -158,7 +158,7 @@ __device__ void blockOutRooms(byte* map, uint map_w, uint map_h, ulong targetCol
 		bool foundEnd = false;
 		while (!foundEnd && endX < map_w)
 		{
-			ulong c = getPixelColor(map, map_w, endX, startY);
+			uint c = getPixelColor(map, map_w, endX, startY);
 			if (c == COLOR_BLACK)
 			{
 				endX += 1;
@@ -174,7 +174,7 @@ __device__ void blockOutRooms(byte* map, uint map_w, uint map_h, ulong targetCol
 		foundEnd = false;
 		while (!foundEnd && endY < map_h)
 		{
-			ulong c = getPixelColor(map, map_w, startX, endY);
+			uint c = getPixelColor(map, map_w, startX, endY);
 			if (c == COLOR_BLACK)
 			{
 				endY += 1;
